messaging_app/requirements.txt
asgiref==3.8.1
django==4.2.25
djangorestframework==3.15.2
sqlparse==0.5.3
typing-extensions==4.13.2
